wip : asset security commitment
